+++ Changelog.md
@@ -1,4 +1,7 @@
 Change log:
+* 2023-05-18 - U35.1.3 - 1 new + 1 newly available + 10 future
+  * Pure Seagull, Cave Slug
+  * Frost Aurochs, Buffalo Cow, Black Bull, Young Moose, Autumn Aurochs, Curious Slug, Feisty Wolverine, Spiky Avanc, Tawny Aurochs, Slate Aurochs
 * 2023-05-02 - U35.1.1 - 4 new + 1 newly available
   * Anniversary Celebration Goat, Rust Corgi, Homesteader's Draught Horse, Dappled Grey Draught Horse, Black Draught Horse
 * 2023-03-23 - U35 - 2 new + 2 newly available